Don't include user mention for anonymous form submissions
We currently use WEBHOOK_ENABLED to determine whether user data should be stored to the db. However, when webhooking a form submission this config is ignored, and the user mention is always included if available.

This means that if a user login in using another form, and then submits an anonymous form with the same session, their name will be included in the webhook.