Allow Setup For Unittests
Adds the setup method as a special method for unittest question types.
Additionally, a couple QoL fixes are included. I changed `unit` to
`self` to clarify what it actually is, and I changed an `lstrip` to a
`removeprefix` to avoid removing unintentional portions of the name
(for instance `something` -> `omething`)

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>