Handle Improper Output For Unittests
The old testing code only checks the first character of the stdout,
meaning that if a user succeeds in printing a 1 before the rest of the
output, the testing code will consider that a pass and discard the rest.

This change checks that the rest of the stdout is empty as expected.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>